+++ src/4uqi/uqi.cc
@@ -79,7 +79,7 @@ uqi_result_get_record(uqi_result_t *result, uint32_t row, ups_record_t *record)
 }
 
 UPS_EXPORT void *UPS_CALLCONV
-uqi_result_get_key_data(uqi_result_t *result, uint64_t *psize)
+uqi_result_get_key_data(uqi_result_t *result, uint32_t *psize)
 {
   Result *r = (Result *)result;
   if (psize)
@@ -88,7 +88,7 @@ uqi_result_get_key_data(uqi_result_t *result, uint64_t *psize)
 }
 
 UPS_EXPORT void *UPS_CALLCONV
-uqi_result_get_record_data(uqi_result_t *result, uint64_t *psize)
+uqi_result_get_record_data(uqi_result_t *result, uint32_t *psize)
 {
   Result *r = (Result *)result;
   if (psize)
